% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvcalc.R
\name{indentPrint}
\alias{indentPrint}
\title{Print with Line Indentation}
\usage{
indentPrint(object, indent = 4, ...)
}
\arguments{
\item{object}{any printable object}

\item{indent}{a non-negative integer, the number of spaces to insert}

\item{\dots}{other arguments to pass to \code{\link{print}}}
}
\value{
\code{object} is returned invisibly
}
\description{
Same as \code{\link{print}}, but adds a specified amount of white space at
the start of each printed line
}
\examples{

indentPrint("this indented by 10 spaces", indent=10)

}
\author{
David Firth, \email{d.firth@warwick.ac.uk}
}
\keyword{IO}
